move year plans state to main page
