+++ site/src/pages/RoadmapPage/PlannerPage.jsx
@@ -4,28 +4,14 @@ import Header from "./Header.jsx";
 import AddYearPopup from "./AddYearPopup.jsx";
 import YearModule from "./YearModule";
 
-function PlannerPage() {
+function PlannerPage({ yearPlans, handleAddYear, removeYear }) {
   const [courseCount, setCourseCount] = useState(0);
   const [unitCount, setUnitCount] = useState(0);
   const [popUp, setPopUp] = useState(false);
-  const [yearPlans, setYearPlans] = useState([]);
 
-  const handleAddYear = (year) => {
-    const newIndex =
-      yearPlans.length !== 0 ? yearPlans[yearPlans.length - 1].index + 1 : 1;
-    const newYear = {
-      index: newIndex,
-      startYear: parseInt(year),
-      courses: 0,
-      units: 0,
-    };
-    setYearPlans([...yearPlans, newYear]);
+  const addYearToPlanner = (year) => {
+    handleAddYear(year);
     setPopUp(false);
-  };
-
-  const removeYear = (year) => {
-    const filteredPlans = yearPlans.filter((plan) => plan.index !== year);
-    setYearPlans(filteredPlans);
   };
 
   return (
@@ -39,7 +25,7 @@ function PlannerPage() {
         })}
       </section>
       <AddYearPopup
-        handleAddYear={handleAddYear}
+        addYearToPlanner={addYearToPlanner}
         setPopUp={setPopUp}
         popUp={popUp}
       />
